Invalidating the session tokens
The exit and logout commands were not properly invalidating the session token, which caused the "CONNECTED" message on subsequent connections. It's now fixed so that logging out properly revokes the session token.
The last Patch was the Fix regarding network Disconnections or just when users accidentally pressed [ CTRL + C ] and logged themselfs out.